requirements.txt
flask
requests
pinecone
google-generativeai